requirements.txt
absl-py==1.0.0
altgraph==0.17.2
astunparse==1.6.3
auto-py-to-exe==2.13.0
bottle==0.12.19
bottle-websocket==0.2.9
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
cycler==0.11.0
Eel==0.12.4
flatbuffers==2.0
fonttools==4.28.5
future==0.18.2
gast==0.5.3
gevent==21.12.0
gevent-websocket==0.10.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.1
numpy==1.21.5
oauthlib==3.2.0
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pefile==2021.9.3
Pillow==8.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
pyparsing==3.0.6
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
pytesseract==0.3.8
python-dateutil==2.8.2
python-dotenv==0.19.2
pywin32-ctypes==0.2.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.3
whichcraft==0.6.1
wrapt==1.13.3
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
